+++ async/src/main/kotlin/ktx/async/async.kt
@@ -7,6 +7,7 @@ import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.suspendCancellableCoroutine
 import kotlinx.coroutines.withContext
+import kotlinx.coroutines.SupervisorJob
 import kotlin.coroutines.ContinuationInterceptor
 import kotlin.coroutines.resume
 
@@ -31,6 +32,18 @@ object KtxAsync : CoroutineScope {
 @Suppress("unused")
 val Dispatchers.KTX
   get() = MainDispatcher
+
+/**
+ * Creates a coroutine scope in the rendering thread with a supervisor job. Allows to manage multiple
+ * tasks executed on the main rendering thread within a single scope, providing mass actions such
+ * as task cancelling that do not affect other scopes.
+ *
+ * An alternative to direct usage of the global [KtxAsync].
+ *
+ * @see kotlinx.coroutines.MainScope
+ */
+@Suppress("FunctionName")
+fun RenderingScope() = CoroutineScope(SupervisorJob() + MainDispatcher)
 
 /**
  * Creates a new [AsyncExecutorDispatcher] wrapping around an [AsyncExecutor] with a single thread to execute tasks