+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 #### 1.9.10-SNAPSHOT
 
+- **[FEATURE]** (`ktx-async`) Added `RenderingScope` factory function for custom scopes using rendering thread dispatcher.
+
 #### 1.9.10-b4
 
 - **[FEATURE]** (`ktx-actors`) Added `onTouchDown`, `onTouchUp` and `onTouchEvent` extension methods that allow to attach `ClickListener` instances to actors.